x/will/types/message_create_wills_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"willstorage/testutil/sample"
)

func TestMsgCreateWills_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateWills
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateWills{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateWills{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
